requirements.txt
Django
django-fernet-fields
django-rosetta
django-statici18n
django-summernote
ipaddress
Pillow
psycopg2
requests
tox
urllib3
